// pruebaGC/src/app/models/computer.ts
export interface Computer {
  _id: string;
  gce_nombre_equipo: string;
  gce_board: string;
  gce_case: string;
  gce_procesador: string;
  gce_grafica: string;
  gce_ram: string;
  gce_disco_duro: string;
  gce_teclado: string;
  gce_mouse: string;
  gce_pantalla: string;
  gce_estado: boolean;
}